style: switch cast tip icon to outline snap glyph
